Add files via upload
This is an updated version of the previous script. This script now operates using a nested dictionary. This will help significantly with the inclusion of more subfolders in the future. As of right now subfolders "CSV" in "Datasets" and "Python" in "Scripts" exist. A folder named "Other" is created for all other files that are in the folder.